+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this repository will be documented in this file.
 
 ### Fixed
 - Fix predict dataset for MQA model
+- Add new loss type - class weights are computed based on adjusted inverse frequency
 - Fix exact match metric displaying 100% when no annotations are present
 
 ## [0.3.0] - 2022-02-07